requirement.txt
boto==2.48.0
certifi==2018.4.16
chardet==3.0.4
Django==1.5.12
django-extensions==2.0.7
django-pagination==1.0.7
django-pagination-bootstrap==1.0.7
django-s3-cache==1.3
django-storages==1.1.8
django-tastypie==0.9.14
djangorestframework==2.3.13
djangotoolbox==1.6.2
idna==2.6
mimeparse==0.1.3
PyMySQL==0.8.1
python-dateutil==2.7.3
python-mimeparse==1.6.0
requests==2.18.4
six==1.11.0
South==1.0
typing==3.6.4
urllib3==1.22
wheel==0.24.0